⬆️ Upgrade for Rust 1.61 Stable + Nightly
This change solves the following issues:

1. Nightly can now build
2. Coverage uses stable flags
3. Bump to 0.2.0-dev (to be expanded to a per-branch/merged to main
   auto-release system)
4. Stabilizes several `unstable` APIs that were stabilized in Rust 1.59
5. Stabilizes the `Termination` trait implementation from `nightly`.
