+++ .github/workflows/coverage.yml
@@ -66,11 +66,10 @@ jobs:
           LLVM_PROFILE_FILE: target/coverage/outcome-%p-%m.profraw
           RUSTC_WRAPPER: sccache
           RUSTDOCFLAGS: >-
-            -Zinstrument-coverage
-            -Zprofile
+            -C instrument-coverage
             -Zunstable-options
             --persist-doctests ${{github.workspace}}/target/coverage
-          RUSTFLAGS: -Zinstrument-coverage -Zprofile
+          RUSTFLAGS: -C instrument-coverage
       - name: Collect Coverage
         run: >-
           grcov ${{github.workspace}}
