+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "outcome-46f94afc-026f-5511-9d7e-7d1fd495fb5c"
 description = "Augmentations for error propagation"
 repository = "https://github.com/bruxisma/outcome"
-version = "0.1.0"
+version = "0.2.0-dev"
 edition = "2021"
 license = "MIT"
 readme = "README.md"
