+++ src/unstable.rs
@@ -2,82 +2,6 @@ use core::cmp::PartialEq;
 
 use crate::prelude::*;
 
-impl<S: Copy, M, F> Outcome<&S, M, F> {
-  /// Maps an `Outcome<&S, M, F>` to an `Outcome<S, M, F>` by copying the
-  /// contents of the `Success` value.
-  ///
-  /// # Examples
-  ///
-  /// ```
-  /// # use outcome::prelude::*;
-  /// let value = 47;
-  /// let x: Outcome<&i32, i32, i32> = Success(&value);
-  /// assert_eq!(x, Success(&47));
-  /// let copied = x.copied();
-  /// assert_eq!(copied, Success(47));
-  /// ```
-  pub fn copied(self) -> Outcome<S, M, F> {
-    self.map(|&s| s)
-  }
-}
-
-impl<S: Copy, M, F> Outcome<&mut S, M, F> {
-  /// Maps an `Outcome<&mut S, M, F>` to an `Outcome<S, M, F>` by copying the
-  /// contents of the `Success` value.
-  ///
-  /// # Examples
-  ///
-  /// ```
-  /// # use outcome::prelude::*;
-  /// let mut value = 47;
-  /// let x: Outcome<&mut i32, i32, i32> = Success(&mut value);
-  /// assert_eq!(x, Success(&mut 47));
-  /// let copied = x.copied();
-  /// assert_eq!(copied, Success(47));
-  /// ```
-  pub fn copied(self) -> Outcome<S, M, F> {
-    self.map(|&mut s| s)
-  }
-}
-
-impl<S: Clone, M, F> Outcome<&S, M, F> {
-  /// Maps an `Outcome<&S, M, F>` to an `Outcome<S, M, F>` by cloning the
-  /// contents of the `Success` value.
-  ///
-  /// # Examples
-  ///
-  /// ```
-  /// # use outcome::prelude::*;
-  /// let val = 47;
-  /// let x: Outcome<&i32, u32, f32> = Success(&val);
-  /// assert_eq!(x, Success(&47));
-  /// let cloned = x.cloned();
-  /// assert_eq!(cloned, Success(47));
-  /// ```
-  pub fn cloned(self) -> Outcome<S, M, F> {
-    self.map(Clone::clone)
-  }
-}
-
-impl<S: Clone, M, F> Outcome<&mut S, M, F> {
-  /// Maps an `Outcome<&mut S, M, F>` to an `Outcome<S, M, F>` by cloning the
-  /// contents of the `Success` value.
-  ///
-  /// # Examples
-  ///
-  /// ```
-  /// # use outcome::prelude::*;
-  /// let mut val = 47;
-  /// let x: Outcome<&mut i32, u32, i32> = Success(&mut val);
-  /// assert_eq!(x, Success(&mut 47));
-  /// let cloned = x.cloned();
-  /// assert_eq!(cloned, Success(47));
-  /// ```
-  pub fn cloned(self) -> Outcome<S, M, F> {
-    self.map(|s| s.clone())
-  }
-}
-
 impl<S, M, F> Outcome<S, M, F> {
   /// Returns `true` if the outcome is a [`Success`] value containing the given
   /// value.
